Fix import of background code in VersionUpgrade component
refactor version.ts into background and client files
